fix: Resolve all clippy warnings and improve code quality
- Fix unused variable `state` in aipub_handler (prefix with _)
- Suppress dead_code for utility functions kept for future use
  (status_i16_to_string, update_status_and_freeze, wallet_repo)
- Add #[allow(clippy::too_many_arguments)] for query filter functions
- Add #[allow(clippy::type_complexity)] for test model tuples
- Fix unused variable in material_integration_test
- Auto-fix needless_borrows_for_generic_args in batch test
- Fix social_account create_account to use dto.platform_id (was hardcoded to 1)
- Add platform_id field to CreateSocialAccountDto with default=2 (TikTok)

+++ crates/api/src/service/social_account_service.rs
@@ -232,7 +232,7 @@ impl SocialAccountService {
         // Parse profile range
         let (prefix, start_num, end_num) = self.parse_profile_range(&dto.profile_start, &dto.profile_end)?;
         
-        let total_count = (end_num - start_num + 1) as i32;
+        let total_count = end_num - start_num + 1;
         
         // Validate: max 100 accounts per batch
         if total_count > 100 {
